refactor event dispatcher
